refactor: store definition IDs in AST instead of objects
The parser now records definitions as arena indices rather than cloning
full definition structures. All downstream consumers (alignment validation,
symbol table building, etc.) have been updated to resolve definitions
from the arena when needed. This reduces memory overhead and ensures
a single source of truth for definition data.

+++ crates/hwc-export/src/bom.rs
@@ -161,7 +161,7 @@ pub fn export(
         bom.push_str("Material,Total_Area_nm2,Total_Volume_nm3,Layer_Count,Coverage_Percentage\n");
 
         // Calculate die area for coverage percentage
-        let die_area_nm2 = space.dimensions.width_nm as i64 * space.dimensions.height_nm as i64;
+        let die_area_nm2 = space.dimensions.width_nm * space.dimensions.height_nm;
 
         // Sort by material name for consistent output
         let mut materials: Vec<_> = material_totals.iter().collect();
